// workpool/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.13.3-alpha.0.
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as client from "../client.js";
import type * as counter from "../counter.js";
import type * as sched from "../sched.js";

import type { ApiFromModules, FunctionReference } from "convex/server";
/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = functions.myModule.myFunction;
 * ```
 */
declare const functions: ApiFromModules<{
  client: typeof client;
  counter: typeof counter;
  sched: typeof sched;
}>;